Create ceaser-cipher.py
Ceaser cipher is a way to encode your message into some code. This code will only be changed to normal message when you decode it. This way no one can really understand what is the meaning of the encoded message. Condition is you should use the same code for encoding and decoding your message. This code uses simplest and most basic functionalities in Python. If else statement and Control loops only.